backend/requirements.txt
django
django-summernote
django-otp
qrcode
django-admin-honeypot
djangorestframework
django-cors-headers
